Factor out PriorityAssemblyResolver
Change to use `TeamFoundationResolver` import attribute with `IResolver` to signal that the import is special.
