+++ src/GitHub.VisualStudio/TeamFoundation/IResolver.cs
@@ -1,0 +1,7 @@
+﻿namespace GitHub.VisualStudio.TeamFoundation
+{
+    public interface IResolver
+    {
+        void Dispose();
+    }
+}
